requirements.txt
blinker
click
colorama
Flask
greenlet
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
pip
protobuf
python-dotenv
setuptools
SQLAlchemy
typing_extensions
Werkzeug
